Conflicts:
	DataReader/HTKMLFReader/ssematrix.h
	Math/Math/CUDAPageLockedMemAllocator.cpp
	Math/Math/Matrix.cpp
	Math/Math/NoGPU.cpp
	Math/Math/ValueQuantizer.h

+++ DataReader/HTKMLFReader/minibatchiterator.h
@@ -209,7 +209,7 @@ public:
     {
         firstvalidepochstartframe = source.firstvalidglobalts (epochstartframe); // epochstartframe may fall between utterance boundaries; this gets us the first valid boundary
         fprintf (stderr, "minibatchiterator: epoch %d: frames [%d..%d] (first utterance at frame %d), data subset %d of %d, with %d datapasses\n",
-                 epoch, epochstartframe, epochendframe, firstvalidepochstartframe, subsetnum, numsubsets, datapasses);
+                 (int)epoch, (int)epochstartframe, (int)epochendframe, (int)firstvalidepochstartframe, (int)subsetnum, (int)numsubsets, (int)datapasses);
         mbstartframe = firstvalidepochstartframe;
         datapass = 0;
         fillorclear(); // get the first batch
@@ -228,7 +228,7 @@ public:
     {
         firstvalidepochstartframe = source.firstvalidglobalts (epochstartframe); // epochstartframe may fall between utterance boundaries; this gets us the first valid boundary
         fprintf (stderr, "minibatchiterator: epoch %d: frames [%d..%d] (first utterance at frame %d), data subset %d of %d, with %d datapasses\n",
-                 epoch, epochstartframe, epochendframe, firstvalidepochstartframe, subsetnum, numsubsets, datapasses);
+                 (int)epoch, (int)epochstartframe, (int)epochendframe, (int)firstvalidepochstartframe, (int)subsetnum, (int)numsubsets, (int)datapasses);
         mbstartframe = firstvalidepochstartframe;
         datapass = 0;
         fillorclear(); // get the first batch
@@ -253,7 +253,7 @@ public:
         {
             mbstartframe = firstvalidepochstartframe;
             datapass++;
-            fprintf (stderr, "\nminibatchiterator: entering %d-th repeat pass through the data\n", datapass+1);
+            fprintf (stderr, "\nminibatchiterator: entering %d-th repeat pass through the data\n", (int)(datapass+1));
         }
         fillorclear();
     }
